+++ results/v3-128/bench/chainhash128_v3.cpp
@@ -1,0 +1,3 @@
+#define REG chainhash128_v3
+#define USE_PUBLIC 1
+#include "registration.inc"